# Add new `Performance/RedundantFetchBlock` cop