--- a/docs/modules/ROOT/pages/cops_performance.adoc
+++ b/docs/modules/ROOT/pages/cops_performance.adoc
@@ -948,6 +948,59 @@ end
 
 * https://github.com/JuanitoFatas/fast-ruby#proccall-and-block-arguments-vs-yieldcode
 
+== Performance/RedundantFetchBlock
+
+|===
+| Enabled by default | Safe | Supports autocorrection | VersionAdded | VersionChanged
+
+| Enabled
+| Yes
+| Yes
+| 1.7
+| -
+|===
+
+This cop identifies places where `fetch(key) { value }`
+can be replaced by `fetch(key, value)`.
+
+In such cases `fetch(key, value)` method is faster
+than `fetch(key) { value }`.
+
+=== Examples
+
+[source,ruby]
+----
+# bad
+hash.fetch(:key) { 5 }
+hash.fetch(:key) { true }
+hash.fetch(:key) { nil }
+array.fetch(5) { :value }
+ENV.fetch(:key) { VALUE }
+ENV.fetch(:key) { 'value' }
+
+# good
+hash.fetch(:key, 5)
+hash.fetch(:key, true)
+hash.fetch(:key, nil)
+array.fetch(5, :value)
+ENV.fetch(:key, VALUE)
+ENV.fetch(:key, 'value')
+----
+
+=== Configurable attributes
+
+|===
+| Name | Default value | Configurable values
+
+| SafeForConstants
+| `false`
+| Boolean
+|===
+
+=== References
+
+* https://github.com/JuanitoFatas/fast-ruby#hashfetch-with-argument-vs-hashfetch--block-code
+
 == Performance/RedundantMatch
 
 |===