@marcandre 2020-06-04
Good PR.

1) Could you please add `true`, `false` (even `nil` although that would be weird).
2) Could you please add a test for symbols with interpolation (no offense)
3) I feel constant strings (without interpolation) should also be passed as arguments; modern If I'm not mistaken ruby runs with shared frozen string constants so there wouldn't be a time cost.
4) Not sure what to do with constants; they can not be completely safely exchanged though. E.g.:

```ruby
h = {a: 1}
h.fetch(:a) { NotDefinedYet } # => 1
NotDefinedYet = 42
h.fetch(:b) { NotDefinedYet } # => NotDefinedYet
```

Autocorrecting the above will fail. You could mark this cop as `Safe: false`, or introduce a setting... Anyone?