@fatkodima 2020-06-04
@marcandre Thanks for review!

1. Oops, I completely forgot about true/false/nil 😄 
Seems like rationals/complex/ranges cannot be used here, as they allocate new instances every time they are used as argument for `fetch` and it is faster to allocate them in a block. Local benchmarks proved that.
2. added
3. same as for point 1
4. added a setting (not sure about a name, maybe there is a better one)