@marcandre 2020-06-04
Right, obviously you're right about Ranges & Complex, my bad.

`Float`s require allocation [only on 32 bit machines](https://bugs.ruby-lang.org/issues/6763) so indeed your mileage may vary.

3) Which Ruby are you testing on? I get

```
Comparison:
  fast hash_with_key:  9353669.7 i/s
  slow hash_with_key:  8974184.9 i/s - same-ish: difference falls within error
```