@fatkodima 2020-06-04
3. Yeah, when the key is present in the hash, there is no difference in performance. 
But there is a difference when the key is missing:
```
Comparison:
fast hash_without_key: 10288181.0 i/s
slow hash_without_key:  6811963.9 i/s - 1.51x  (± 0.00) slower
```

So seems like the case when block returns a String isn't worth considering.

Tested on `ruby 2.6.5p114`