@marcandre 2020-06-05
> But there is a difference when the key is missing

Yes, that's the point, right, the block form will always best equal or slower when the key is missing. It's thus best to avoid the block form for strings.