@fatkodima 2020-06-05
@marcandre Yes, sure, you are correct! I mixed up `fast` and `slow` parts in my benchmark and didn't notice this. I wondered after your first comment, why block with string is faster than string as argument when strings are frozen but trusted the benchmark.

Updated to check for strings.

Thanks for patience! 😄