@fatkodima 2020-06-11
> 1. Could you please add true, false (even nil although that would be weird), rationals and complex numbers too, constant Ranges. I think that immutable_litteral? or basic_litteral? should be a good base.

You were correct regarding `Rational`s and `Complex`s, I was not testing properly and forgot about that there are literals for them.

Updated with usage of `basic_literal?` + `const_type?`.