@bbatsov 2020-06-12
That's one of those cases that sit between style and performance. For me it's a given that you shouldn't be using a block unless you have to compute something, regardless of the performance implications.