@fatkodima 2020-06-12
@bbatsov So, wdyt, is this better to be moved to `Style` in main repo?