@bbatsov 2020-06-12
I'm on the fence. @rubocop-hq/rubocop-core what do you think?