@koic 2020-06-12
Yeah, I think this cop can move to `Style` department.