@fatkodima 2020-06-12
Transferred https://github.com/rubocop-hq/rubocop/pull/8147